EditPuzzleCell file & vars renamed
