+++ web/src/pages/EditPuzzleFormPage/EditPuzzleFormPage.tsx
@@ -1,11 +1,11 @@
-import EditPuzzleCell from 'src/components/EditPuzzleCell'
+import EditRewardableCell from 'src/components/EditRewardableCell'
 import Seo from 'src/components/Seo/Seo'
 
 const EditPuzzleFormPage = ({ slug }: { slug: string }) => {
   return (
     <>
       <Seo title="Edit your puzzle" />
-      <EditPuzzleCell slug={slug} />
+      <EditRewardableCell slug={slug} />
     </>
   )
 }
